// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  caption?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    hero?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
// ...
export interface News {
  id: number;
  title: string;
  slug: string;
  publishedDate: string;
  featuredImage?: (number | null) | Media;
  /**
   * Breve descrizione per anteprime (max 200 caratteri)
   */
  excerpt?: string | null;
  content: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  tags?: ('bioedilizia' | 'sostenibilita' | 'innovazione' | 'progetti' | 'eventi' | 'normative')[] | null;
  seo?: {
    metaTitle?: string | null;
    metaDescription?: string | null;
  };
  status?: ('draft' | 'published') | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}